refactor(finances): extract FacilitiesCard and PayrollTable sub-components
The facilities grid and payroll table are lifted out of FinancesTab.tsx
into FinancesFacilitiesCard.tsx and FinancesPayrollTable.tsx. Both receive
data and callbacks as props; all state and handlers remain in the parent.
FinancesTab.tsx drops another ~150 lines (~980 lines total, down from 1330).
